extracted two test to outmost level
